use strict formatting for ollama
