+++ src/utils/text.js
@@ -25,3 +25,38 @@ export function toSinglePrompt(turns, system=null, stop_seq='***', model_nicknam
         prompt += model_nickname + ": ";
     return prompt;
 }
+
+// ensures stricter turn order for anthropic/llama models
+// combines repeated messages from the same role, separates repeat assistant messages with filler user messages
+export function strictFormat(turns) {
+    let prev_role = null;
+    let messages = [];
+    let filler = {role: 'user', content: '_'};
+    for (let msg of turns) {
+        if (msg.role === 'system') {
+            msg.role = 'user';
+            msg.content = 'SYSTEM: ' + msg.content;
+        }
+        if (msg.role === prev_role && msg.role === 'assistant') {
+            // insert empty user message to separate assistant messages
+            messages.push(filler);
+            messages.push(msg);
+        }
+        else if (msg.role === prev_role) {
+            // combine new message with previous message instead of adding a new one
+            messages[messages.length-1].content += '\n' + msg.content;
+        }
+        else {
+            messages.push(msg);
+        }
+        prev_role = msg.role;
+        
+    }
+    if (messages.length > 0 && messages[0].role !== 'user') {
+        messages.unshift(filler); // anthropic requires user message to start
+    }
+    if (messages.length === 0) {
+        messages.push(filler);
+    }
+    return messages;
+}
